+++ client/lib/plugins/color-functions/color-functions.js
@@ -4,7 +4,7 @@
 // Copyright (c) 2018 Alexandre Bento Freire. All rights reserved.
 // Licensed under the MIT License+uuid License. See License.txt for details
 // ------------------------------------------------------------------------
-// Implements a list of built-in text Tasks
+// Implements a list of color functions
 /** @module end-user | The lines bellow convey information for the end-user */
 /**
  * ## Description